[sailfishkeyprovider] Have more specific warning when key not found. Contributes to JB#50147
"reached end of file" is quite a vague warning on the journal when
trying to fetch a key that doesn't exist. Could be considered callers
responsibility to note that to journal, but having an exact error
should be better too, and helps detecting if there are wrong assumptions
on file content.